% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggirl.R
\name{address}
\alias{address}
\title{Create an address object}
\usage{
address(
  name,
  address_line_1,
  address_line_2 = NULL,
  city,
  state = NULL,
  postal_code,
  country
)
}
\arguments{
\item{name}{The name for the address}

\item{address_line_1}{The first line of the address}

\item{address_line_2}{(Optional) A second address line, such as an apartment number.}

\item{city}{the city}

\item{state}{(Optional) The state to send to}

\item{postal_code}{The postal code (ZIP code in the US)}

\item{country}{The 2-character \href{https://en.wikipedia.org/wiki/List_of_ISO_3166_country_codes}{ISO-1366 code} for the country. Non-US shipping is experimental!}
}
\description{
This function takes string inputs and converts them into an address object that can be used to send irl art (or as a return address).
}
\examples{
send_address <- address(name = "RStudio", address_line_1 = "250 Northern Ave", city = "Boston", state = "MA", postal_code = "02210", country = "US")

}
