% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggartprint.R
\name{ggartprint}
\alias{ggartprint}
\title{Order art prints of your ggplot!}
\usage{
ggartprint(
  plot,
  size = "11x14",
  orientation = c("landscape", "portrait"),
  quantity = 1,
  contact_email,
  address,
  ...
)
}
\arguments{
\item{plot}{the plot to use as an art print.}

\item{size}{the size of the art print. Use \code{\link[=ggartprint_sizes]{ggartprint_sizes()}} to see a list of the sizes. If a size isn't available that you want email support@ggirl.art for custom sizes.}

\item{orientation}{should the plot be landscape or portrait?}

\item{quantity}{the number of prints to order (defaults to 1).}

\item{contact_email}{email address to send order updates.}

\item{address}{the physical address to mail the print(s) to. Use the \code{\link[=address]{address()}} function to format it.}

\item{...}{other options to pass to ragg::agg_png when turning the plot into an image for the front of the postcard.}
}
\description{
This function takes a ggplot2 output and will order an art print to hang on a wall!
Running this function will bring you to a webpage to confirm the order and submit it.
\emph{No order will be submitted until you explicitly approve it.}
}
\details{
You can choose from a number of options for the size of the print (and either rectangular or square).
All of the sizes are high resolution, so things like text size in the R/RStudio plot may not reflect what
it would look like as a poster. It's recommended you run the function a few times and adjust plot attributes
until you get it the way you like it.

Prints take up to 3-4 weeks to deliver.
}
\examples{
library(ggplot2)
library(ggirl)
delivery_address <- address(name = "Fake person", address_line_1 = "101 12th st",
  address_line_2 = "Apt 17", city = "Seattle", state = "WA",
  postal_code = "98102", country = "US")
contact_email = "fakeemail275@gmail.com"
plot <- ggplot(data.frame(x=1:10, y=runif(10)),aes(x=x,y=y))+geom_line()+geom_point()+theme_gray(48)
ggartprint(plot, size="24x36", orientation = "landscape", quantity = 1,
           contact_email = contact_email, address = delivery_address)
}
\seealso{
\code{\link[=address]{address()}} to format an address for ggirl
}
