% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggartprint.R
\name{ggartprint_preview}
\alias{ggartprint_preview}
\title{Preview your art print}
\usage{
ggartprint_preview(plot, size, orientation, ...)
}
\arguments{
\item{plot}{the plot to use as an art print}

\item{size}{the size of the art print. Use \code{\link[=ggartprint_sizes]{ggartprint_sizes()}} to see a list of the sizes. If a size isn't available that you want email support@ggirl.art for custom sizes.}

\item{orientation}{should the plot be landscape or portrait?}

\item{...}{other options to pass to \code{ragg::agg_png()} when turning the plot into an image for the front of the postcard.}
}
\description{
This function takes a ggplot2 output and gives a preview of how the plot will look as an art print.
While it's totally fine to just call ggirl::ggartprint to preview, this allows you to preview locally.
}
\details{
The preview will appear in either the "Viewer" pane of RStudio or in your browser, depending on if RStudio is installed or not.
The preview includes a frame, but that will not be included with the print.
}
\examples{
library(ggplot2)
library(ggirl)
plot <- ggplot(data.frame(x=1:10, y=runif(10)),aes(x=x,y=y))+geom_line()+geom_point()+theme_gray(48)
ggartprint_preview(plot, size="24x36", orientation = "landscape")
}
\seealso{
\code{\link[=ggartprint]{ggartprint()}} to order the art print
}
