% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggartprint.R
\name{ggartprint_sizes}
\alias{ggartprint_sizes}
\title{get a table of sizes of prints available.}
\usage{
ggartprint_sizes()
}
\description{
Prices include shipping. If a size isn't available that you want email support@ggirl.art for custom sizes.
}
