% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpostcard.R
\name{ggpostcard}
\alias{ggpostcard}
\title{Order postcards of your ggplot!}
\usage{
ggpostcard(
  plot = last_plot(),
  contact_email,
  messages,
  send_addresses,
  return_address = NULL,
  ...
)
}
\arguments{
\item{plot}{the plot to put on the front of the postcard}

\item{contact_email}{email address to send order updates}

\item{messages}{either a message to use with all of the recipients, or a list of messages of the same length as the list of addresses (one for each address).}

\item{send_addresses}{either a result of the "address()" function, or a list of results of the "address()" function.}

\item{return_address}{(optional) the return address for the postcard. \strong{Must be a US address.}}

\item{...}{other options to pass to ggsave when turning the plot into an image for the front of the postcard}
}
\description{
This function takes a ggplot2 output and will send postcards of it for you!
Running this function will bring you to a webpage to confirm the order and submit it.
\emph{No order will be submitted until you explicitly approve it.}
}
\examples{
library(ggplot2)
library(ggirl)
return_address <- address(name = "Jacqueline Nolis", address_line_1 = "111 North St",
                          city = "Seattle", state = "WA",
                          postal_code = "11111", country = "US")
contact_email <- "fakeemailforreal@gmail.com"
send_address_1 <- address(name = "Fake Personname", address_line_1 = "250 North Ave",
                          city = "Boston", state = "MA",
                          postal_code = "22222", country = "US")
send_address_2 <- address(name = "Anotherfake personname",
                          address_line_1 = "15 NE 36th St", address_line_2 = "Apt 4",
                          city = "Redmond", state = "WA",
                          postal_code = "33333", country = "US")

plot <- ggplot(data.frame(x=1:10, y=runif(10)),aes(x=x,y=y))+geom_line()+geom_point()

# send to one recipient
ggpostcard(plot, contact_email,  messages = "An example postcard", send_addresses = send_address_1)

# send the same message to multiple recipients
ggpostcard(plot, contact_email, messages = "An example postcard", send_addresses = list(send_address_1, send_address_2))

# send different messages to multiple recipients, and a return address
ggpostcard(plot, contact_email,
           messages = c("message for sender 1","message for sender 2"),
           send_addresses = list(send_address_1, send_address_2),
           return_address = return_address)
}
\seealso{
\code{\link[=ggpostcard_preview]{ggpostcard_preview()}} to preview a plot on a postcard within R

\code{\link[=address]{address()}} to format an address for ggirl
}
