% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpostcard_example_contouR.R
\name{ggpostcard_example_contouR}
\alias{ggpostcard_example_contouR}
\title{An example postcard with art from the contouR package}
\usage{
ggpostcard_example_contouR(background_col = NULL, line_col = NULL)
}
\arguments{
\item{background_col}{(optional) the background color of the plot}

\item{line_col}{(optional) the color of the lines in the plot}
}
\value{
a ggplot2 plot to pass to ggpostcard
}
\description{
This example makes a postcard with a beautiful graphic using the \href{https://github.com/Ijeamakaanyene/contouR}{contouR package} by Ijeamaka Anyene.
}
\examples{
library(ggirl)
return_address <- address(name = "Jacqueline Nolis", address_line_1 = "111 North St",
                          city = "Seattle", state = "WA",
                          postal_code = "11111", country = "US")
contact_email <- "fakeemailforreal@gmail.com"
send_addresses <- address(name = "Fake Personname", address_line_1 = "250 North Ave",
                          city = "Boston", state = "MA",
                          postal_code = "22222", country = "US")
messages <- "Look at this cool plot I found!"
plot <- ggpostcard_example_contouR()
ggpostcard(plot = plot, contact_email = contact_email, return_address = return_address,
  send_addresses = send_addresses, messages = messages)
}
\seealso{
Other ggpostcard_examples: 
\code{\link{ggpostcard_example_rstereogram}()},
\code{\link{ggpostcard_example_sunrise}()}
}
\concept{ggpostcard_examples}
