% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpostcard_example_sunrise.R
\name{ggpostcard_example_sunrise}
\alias{ggpostcard_example_sunrise}
\title{An example postcard with the sunrise and sunset times for a location}
\usage{
ggpostcard_example_sunrise(
  location_lat,
  location_long,
  location_tz,
  location_name,
  ...
)
}
\arguments{
\item{location_lat}{the latitude of the location}

\item{location_long}{the longitude of the location}

\item{location_tz}{a tz time zone string of the style taken by R functions like as.POSIXct}

\item{location_name}{string for the name to show for the location (ex: "Seattle, WA")}

\item{...}{other options to pass to ggpostcard()}
}
\description{
This example makes a postcard with a pretty ggplot that gives you when sunrise
and sunset is over the course of the year.
}
\details{
To get the longitude and latitude for your location, you can click a place in Google maps.
To get the time zone, you can use the \href{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones}{Wikipedia list of time zone codes}

This function uses an external API \href{https://sunrise-sunset.org/api}{https://sunrise-sunset.org/api}. While the package has some built-in caching,
_if you call the function with too many locations in a short period of time your IP will be temporarily blocked_.
}
\examples{
library(ggirl)
location_lat <- 47.6062
location_long <- -122.3321
location_tz <- "America/Los_Angeles"
location_name <- "Seattle, WA"
return_address <- address(name = "Jacqueline Nolis", address_line_1 = "111 North St",
                          city = "Seattle", state = "WA",
                          postal_code = "11111", country = "US")
contact_email <- "fakeemailforreal@gmail.com"
send_addresses <- address(name = "Fake Personname", address_line_1 = "250 North Ave",
                          city = "Boston", state = "MA",
                          postal_code = "22222", country = "US")
messages <- "Look at this cool plot I found!"
ggpostcard_example_sunrise(location_lat, location_long, location_tz, location_name,
  contact_email = contact_email, return_address = return_address,
  send_addresses = send_addresses, messages = messages)
}
