% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpostcard.R
\name{ggpostcard_preview}
\alias{ggpostcard_preview}
\title{Preview the front of your postcard}
\usage{
ggpostcard_preview(plot, ...)
}
\arguments{
\item{plot}{the plot to put on the front of the postcard}

\item{...}{other options to pass to ggsave when turning the plot into an image for the front of the postcard}
}
\description{
This function takes a ggplot2 output and gives a preview of how the image will look.
While it's totally fine to just call ggirl::ggpostcard to preview, this allows you to preview before having
the addresses and other details set.
}
\details{
The preview will appear in either the "Viewer" pane of RStudio or in your browser, depending on if RStudio is installed or not
}
\examples{
library(ggplot2)
library(ggirl)
plot <- ggplot(data.frame(x=1:10, y=runif(10)),aes(x=x,y=y))+geom_line()+geom_point()
ggpostcard_preview(plot)
}
