% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwatercolor.R
\name{ggwatercolor}
\alias{ggwatercolor}
\title{Request a watercolor commission of your ggplot!}
\usage{
ggwatercolor(
  plot,
  orientation = c("landscape", "portrait"),
  contact_email,
  address,
  ...
)
}
\arguments{
\item{plot}{the plot to use as an art print.}

\item{orientation}{should the plot be landscape or portrait?}

\item{contact_email}{email address to send order updates.}

\item{address}{the physical address to mail the painting to. Use the \code{\link[=address]{address()}} function to format it.}

\item{...}{other options to pass to \code{ragg::agg_png()} when turning the plot into an image for the front of the postcard.}

\item{quantity}{the number of prints to order (defaults to 1).}
}
\description{
This function takes a ggplot2 output and request a handpainted watercolor painting of it!
Running this function will bring you to a webpage to confirm the request--a followup email will contain the details and how to pay.
\emph{No painting will be made until after an email exchange and the payment sent.}
}
\details{
Watercolor paintings will be made 8"x10" on 140lb cold-press paper with professional light-safe paints.
Since these are painted by hand, they may not be exactly accurate to the original, and may be simplified
depending on the complexity of the original (which will be discussed by email).

The paintings take up to 4 weeks to be delivered.
}
\examples{
library(ggplot2)
library(ggirl)
delivery_address <- address(name = "Fake person", address_line_1 = "101 12th st",
  address_line_2 = "Apt 17", city = "Seattle", state = "WA",
  postal_code = "98102", country = "US")
contact_email = "fakeemail275@gmail.com"
plot <- ggplot(data.frame(x=1:10, y=runif(10)),aes(x=x,y=y))+geom_line()+geom_point()+theme_gray(48)
ggwatercolor(plot, orientation = "landscape",
           contact_email = contact_email, address = delivery_address)
}
\seealso{
\code{\link[=address]{address()}} to format an address for ggirl
}
