% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwatercolor.R
\name{ggwatercolor_preview}
\alias{ggwatercolor_preview}
\title{Preview your watercolor request}
\usage{
ggwatercolor_preview(plot, size, orientation, ...)
}
\arguments{
\item{plot}{the plot to use as an art print}

\item{orientation}{should the plot be landscape or portrait?}

\item{...}{other options to pass to \code{ragg::agg_png()} when turning the plot into an image.}
}
\description{
This function takes a ggplot2 output and gives a preview of how plot will be requested to be painted.
While it's totally fine to just call ggirl::ggwatercolor to preview, this allows you to preview locally.
}
\details{
The preview will appear in either the "Viewer" pane of RStudio or in your browser, depending on if RStudio is installed or not.
The preview includes a frame, but that will not be included with the print.
}
\examples{
library(ggplot2)
library(ggirl)
plot <- ggplot(data.frame(x=1:10, y=runif(10)),aes(x=x,y=y))+geom_line()+geom_point()+theme_gray(48)
ggwatercolor_preview(plot, orientation = "landscape")
}
\seealso{
\code{\link[=ggwatercolor]{ggwatercolor()}} to request the watercolor commission
}
