% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggirl.R
\name{upload_data_and_launch}
\alias{upload_data_and_launch}
\title{Upload data}
\usage{
upload_data_and_launch(data, server_url, type)
}
\description{
This function will upload the data to the server, then launch the page for it
}
